set output to be: Numb_of_Bits, Ex_Time

+++ GCD.c
@@ -29,7 +29,7 @@ unsigned int countBits(unsigned int n)
 int main(int argc, char *argv[]){
     char *err;
     long long int a, b, f;
-    int bits;
+    long long int bits;
 
     a = strtoll(argv[1], &err, 10);
     b = strtoll(argv[2], &err, 10); 
@@ -47,7 +47,7 @@ int main(int argc, char *argv[]){
     clock_t end = clock();  // Records the stop time of the program
     float time = ( double )( end - begin ) / CLOCKS_PER_SEC;  // Calculates the elapsed time in seconds
     
-    printf("%i\t", bits);
+    printf("%lli\t", bits);
     //printf("%lli\t", f);
     printf("%f\n", time);
     return 0;
