+++ GCD_Extendido.c
@@ -54,7 +54,7 @@ unsigned int countBits(unsigned int n)
 int main(int argc, char* argv[]){
     char *err;
     long long int a, b;
-    int bits;
+    long long int bits;
 
     a = strtoll(argv[1], &err, 10);
     b = strtoll(argv[2], &err, 10); 
@@ -72,7 +72,7 @@ int main(int argc, char* argv[]){
     clock_t end = clock();  // Records the stop time of the program
     float time = ( double )( end - begin ) / CLOCKS_PER_SEC;  // Calculates the elapsed time in seconds
 
-    printf("%i\t", bits);
+    printf("%lli\t", bits);
     printf("%f\n", time) ;// Prints out the two numbers separated by a space and followed by a newline
     return 0;
 }
