+++ GCD_Extendido_Bin.c
@@ -67,7 +67,7 @@ unsigned int countBits(unsigned int n)
 int main(int argc, char *argv[]){
     char *err;
     long long int x, y, g, u, v, A, B, C, D, PivoteX, PivoteY;
-    int bits;
+    long long int bits;
 
     g = 1;
     x = strtoll(argv[1], &err, 10);
@@ -101,7 +101,7 @@ int main(int argc, char *argv[]){
     clock_t end = clock();  // Records the stop time of the program
     float time = ( double )( end - begin ) / CLOCKS_PER_SEC;  // Calculates the elapsed time in seconds
 
-    printf("%i\t", bits);
+    printf("%lli\t", bits);
     printf("%f\n", time) ;// Prints out the two numbers separated by a space and followed by a newline
     return 0;
 }
